Simplify the seeder and add some configurations

+++ views/seeder.blade.php
@@ -4,6 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Config;
 
 class LaratrustSeeder extends Seeder
 {
@@ -24,7 +25,7 @@ class LaratrustSeeder extends Seeder
         foreach ($config as $key => $modules) {
 
             // Create a new role
-            $role = \{{ $role }}::create([
+            $role = \{{ $role }}::firstOrCreate([
                 'name' => $key,
                 'display_name' => ucwords(str_replace('_', ' ', $key)),
                 'description' => ucwords(str_replace('_', ' ', $key))
@@ -53,51 +54,17 @@ class LaratrustSeeder extends Seeder
             // Attach all permissions to the role
             $role->permissions()->sync($permissions);
 
-            $this->command->info("Creating '{$key}' user");
-
-            // Create default user for each role
-            $user = \{{ $user }}::create([
-                'name' => ucwords(str_replace('_', ' ', $key)),
-                'email' => $key.'@app.com',
-                'password' => bcrypt('password')
-            ]);
-
-            $user->attachRole($role);
-        }
-
-        // Creating user with permissions
-        if (!empty($userPermission)) {
-
-            foreach ($userPermission as $key => $modules) {
-
-                foreach ($modules as $module => $value) {
-
-                    // Create default user for each permission set
-                    $user = \{{ $user }}::create([
-                        'name' => ucwords(str_replace('_', ' ', $key)),
-                        'email' => $key.'@app.com',
-                        'password' => bcrypt('password'),
-                        'remember_token' => Str::random(10),
-                    ]);
-                    $permissions = [];
-
-                    foreach (explode(',', $value) as $p => $perm) {
-
-                        $permissionValue = $mapPermission->get($perm);
-
-                        $permissions[] = \{{ $permission }}::firstOrCreate([
-                            'name' => $permissionValue . '-' . $module,
-                            'display_name' => ucfirst($permissionValue) . ' ' . ucfirst($module),
-                            'description' => ucfirst($permissionValue) . ' ' . ucfirst($module),
-                        ])->id;
-
-                        $this->command->info('Creating Permission to '.$permissionValue.' for '. $module);
-                    }
-                }
-
-                // Attach all permissions to the user
-                $user->permissions()->sync($permissions);
+            if(Config::get('laratrust_seeder.create_users')) {
+                $this->command->info("Creating '{$key}' user");
+                // Create default user for each role
+                $user = \{{ $user }}::create([
+                    'name' => ucwords(str_replace('_', ' ', $key)),
+                    'email' => $key.'@app.com',
+                    'password' => bcrypt('password')
+                ]);
+                $user->attachRole($role);
             }
+
         }
     }
 
@@ -113,19 +80,23 @@ class LaratrustSeeder extends Seeder
         DB::table('{{ config('laratrust.tables.permission_role') }}')->truncate();
         DB::table('{{ config('laratrust.tables.permission_user') }}')->truncate();
         DB::table('{{ config('laratrust.tables.role_user') }}')->truncate();
-        $usersTable = (new \{{ $user }})->getTable();
         $rolesTable = (new \{{ $role }})->getTable();
         $permissionsTable = (new \{{ $permission }})->getTable();
-        DB::statement("TRUNCATE TABLE {$usersTable} CASCADE");
         DB::statement("TRUNCATE TABLE {$rolesTable} CASCADE");
         DB::statement("TRUNCATE TABLE {$permissionsTable} CASCADE");
+        if(Config::get('laratrust_seeder.truncate_tables')) {
+            $usersTable = (new \{{ $user }})->getTable();
+            DB::statement("TRUNCATE TABLE {$usersTable} CASCADE");
+        }
 @else
         DB::table('{{ config('laratrust.tables.permission_role') }}')->truncate();
         DB::table('{{ config('laratrust.tables.permission_user') }}')->truncate();
         DB::table('{{ config('laratrust.tables.role_user') }}')->truncate();
-        \{{ $user }}::truncate();
         \{{ $role }}::truncate();
         \{{ $permission }}::truncate();
+        if(Config::get('laratrust_seeder.truncate_tables')) {
+        \{{ $user }}::truncate();
+        }
 @endif
         Schema::enableForeignKeyConstraints();
     }
